+++ treeComponentFramework/sweepercraft/SweeperCraft.ts
@@ -79,7 +79,7 @@ class SweeperCraftGame {
   getGameMessage() {
     if (this.isLost()) return "You Lost :("
     else if (this.isWon()) return "You won!"
-    return ""
+    return " "
   }
 
   getGameStateClass() {
@@ -205,14 +205,14 @@ class SweeperCraftGame {
     this._render()
   }
 
-  _setBoard(board: Board) {
+  private _setBoard(board: Board) {
     if (!(board instanceof Array)) throw new Error("Invalid Board. Board must be an Array.")
     if (!board.length) throw new Error("Invalid Board. No rows in Board. Expected: Row[]")
     if (!board[0].length) throw new Error("Invalid Board. No columns in row. Expected Row to be: int[]")
     this._board = board
   }
 
-  _resetBoard() {
+  private _resetBoard() {
     clearInterval(this._replayInterval)
     this._numberOfMines = SweeperCraftGame.sum(this._board)
     this._numberOfRows = this._board.length
@@ -224,19 +224,19 @@ class SweeperCraftGame {
     this._resetState()
   }
 
-  _resetState() {
+  private _resetState() {
     this._state = 0
   }
 
-  _zeroedBoard() {
+  private _zeroedBoard() {
     return SweeperCraftGame.getZeroedBoard(this._numberOfRows, this._numberOfColumns)
   }
 
-  _resetClicked() {
+  private _resetClicked() {
     this._clicked = this._zeroedBoard()
   }
 
-  _click(row: int, column: int) {
+  private _click(row: int, column: int) {
     this._clicked[row][column] = 1
     if (this.hasBomb(row, column)) {
       this._lose()
@@ -249,7 +249,7 @@ class SweeperCraftGame {
     }
   }
 
-  _clickNeighbors(row: int, column: int) {
+  private _clickNeighbors(row: int, column: int) {
     this._getNeighbors(row, column).map(coordinate => {
       const row = coordinate[0]
       const col = coordinate[1]
@@ -268,16 +268,16 @@ class SweeperCraftGame {
     if (this._renderFn) this._renderFn(this)
   }
 
-  _getNeighbors(row: int, column: int) {
+  private _getNeighbors(row: int, column: int) {
     return SweeperCraftGame.getNeighbors(row, column, this._numberOfRows, this._numberOfColumns)
   }
 
-  _win() {
+  private _win() {
     this._endTime = Date.now()
     this._state = 1
   }
 
-  _lose() {
+  private _lose() {
     this._endTime = Date.now()
     this._state = 2
   }
@@ -305,7 +305,7 @@ class SweeperCraftGame {
     const numCols = board[0].length
     const c = board.map(row => row.join("")).join("")
     const strLength = c.length
-    var layout = ""
+    let layout = ""
     for (let i = 0; i < strLength; i = i + 6) {
       layout += SweeperCraftGame._bitsToChar(c.substr(i, 6))
     }
@@ -329,20 +329,20 @@ class SweeperCraftGame {
     }
 
     const layout = options.layout
-    var board = SweeperCraftGame.getZeroedBoard(numRows, numCols)
+    let board = SweeperCraftGame.getZeroedBoard(numRows, numCols)
 
     const expectedSquares = numRows * numCols
-    var boardStr = ""
-    for (let i = 0; i < layout.length; i++) {
-      boardStr += SweeperCraftGame._charToSixBits(layout[i])
+    let boardStr = ""
+    for (let rowIndex = 0; rowIndex < layout.length; rowIndex++) {
+      boardStr += SweeperCraftGame._charToSixBits(layout[rowIndex])
     }
 
     boardStr = boardStr.substr(0, expectedSquares)
     board = []
-    for (var i = 0; i < numRows; i++) {
+    for (let rowIndex = 0; rowIndex < numRows; rowIndex++) {
       board.push(
         boardStr
-          .substr(i * numCols, numCols)
+          .substr(rowIndex * numCols, numCols)
           .split("")
           .map(c => parseInt(c))
       )
@@ -449,8 +449,8 @@ class SweeperCraftGame {
   static _getPermalinkMap() {
     if (!this._permalinkMap) {
       this._permalinkMap = {}
-      this._permalinkArr.forEach((l, i) => {
-        this._permalinkMap[l] = i
+      this._permalinkArr.forEach((letter, index) => {
+        this._permalinkMap[letter] = index
       })
     }
     return this._permalinkMap
@@ -473,7 +473,7 @@ const linkToObject = (link: string): Object => {
   const parts = link.replace(/^\//, "").split("/")
   const obj: any = {}
   const length = parts.length
-  for (var index = 0; index < length; index = index + 2) {
+  for (let index = 0; index < length; index = index + 2) {
     obj[parts[index]] = parts[index + 1]
   }
   return obj
@@ -535,6 +535,7 @@ class SweeperCraftApp extends AbstractTreeComponentRootNode {
       headerComponent: headerComponent,
       boardComponent: boardComponent,
       controlsComponent: controlsComponent,
+      customLinkComponent: customLinkComponent,
       shortcutsTableComponent: shortcutsTableComponent,
       githubTriangleComponent: githubTriangleComponent,
       TreeComponentFrameworkDebuggerComponent: TreeComponentFrameworkDebuggerComponent
@@ -676,6 +677,7 @@ class SweeperCraftApp extends AbstractTreeComponentRootNode {
     return `headerComponent
 boardComponent
 controlsComponent
+customLinkComponent
 shortcutsTableComponent
 githubTriangleComponent`
   }
@@ -725,13 +727,23 @@ githubTriangleComponent`
     }
   }
 
+  // todo: there's probably a better pattern than this.
+  private _syncBoardToGame() {
+    this.getTopDownArray()
+      .filter((node: any) => node instanceof AbstractSweeperCraftComponent)
+      .forEach((node: AbstractSweeperCraftComponent) => {
+        node._syncBoardToGame()
+      })
+  }
+
   private _loadFromHash(stumpNode: any) {
     const link = location.hash.replace(/^\#/, "")
     let board
     if (!link) board = SweeperCraftGame.getRandomBoard(9, 9, 10)
     else board = SweeperCraftGame.boardFromPermalink(link)
     this._mainGame = new SweeperCraftGame(board, game => {
-      this.makeAllDirty() // todo: cleanup
+      this._syncBoardToGame() // todo: cleanup
+
       this.renderAndGetRenderResult(stumpNode)
     })
     let boardNode = this.getNode("boardComponent")
@@ -780,10 +792,11 @@ githubTriangleComponent`
   }
 }
 
-class headerComponent extends AbstractTreeComponent {
-  // mines moves gameTime gameMessage
-  // 10 1 You Lost!
+abstract class AbstractSweeperCraftComponent extends AbstractTreeComponent {
+  abstract _syncBoardToGame(): void
+}
 
+class headerComponent extends AbstractSweeperCraftComponent {
   treeComponentDidMount() {
     super.treeComponentDidMount()
     this._initTimerInterval()
@@ -826,6 +839,12 @@ class headerComponent extends AbstractTreeComponent {
     return this.getRootNode().getGame()
   }
 
+  // mines moves gameMessage
+  // 10 1 You Lost!
+  _syncBoardToGame() {
+    this.setContent(`${this.numberOfMines}mines ${this.numberOfMoves}clicks ${this.gameMessage}`)
+  }
+
   getStumpCode() {
     return `div
  class headerComponent
@@ -845,18 +864,27 @@ class headerComponent extends AbstractTreeComponent {
   }
 }
 
-class boardComponent extends AbstractTreeComponent {
+class boardComponent extends AbstractSweeperCraftComponent {
   createParser() {
     return new jtree.TreeNode.Parser(undefined, {
       rowComponent: rowComponent
     })
   }
 
+  _syncBoardToGame() {
+    this.setContent(`${this._getCssGameClass()}`)
+  }
+
+  _getCssGameClass() {
+    return this.getRootNode()
+      .getGame()
+      .isOver()
+      ? "over"
+      : "playing"
+  }
+
   getCssClassNames() {
-    const game = this.getRootNode().getGame()
-    const rows = game.getBoard()
-    const className = game.isOver() ? "over" : "playing"
-    return `${className} ${super.getCssClassNames()}`
+    return `${this._getCssGameClass()} ${super.getCssClassNames()}`
   }
 }
 
@@ -868,23 +896,38 @@ class rowComponent extends AbstractTreeComponent {
   }
 }
 
-class squareComponent extends AbstractTreeComponent {
+class squareComponent extends AbstractSweeperCraftComponent {
   getStumpCode() {
-    const game = this.getRootNode().getGame()
-    const row = this.getParent().getIndex()
-    const col = this.getIndex()
-    const wasClicked = game.wasClicked(row, col)
-    let content = ""
-    if (wasClicked) {
-      const neighborBombCount = game.getNeighborBombCount(row, col)
-      content = neighborBombCount ? neighborBombCount : " "
-    }
+    const row = this.getRow()
+    const col = this.getColumn()
 
-    return `div${content ? " " + content : ""}
+    return `div${this.htmlContent}
  stumpOnClickCommand clickSquareCommand ${row} ${col}
  stumpOnShiftClickCommand flagSquareCommand ${row} ${col}
  stumpOnContextMenuCommand flagSquareCommand ${row} ${col}
  class ${this.getCssClassNames()}`
+  }
+
+  _syncBoardToGame() {
+    this.setContent(`${this.wasClicked ? "clicked" : "notClicked"} ${this.isFlagged ? "flagged" : "notFlagged"}`)
+  }
+
+  get isFlagged() {
+    return this.game.isFlagged(this.getRow(), this.getColumn())
+  }
+
+  get htmlContent() {
+    if (!this.wasClicked) return ""
+
+    return " " + (this.neighborBombCount || "")
+  }
+
+  get wasClicked() {
+    return this.game.wasClicked(this.getRow(), this.getColumn())
+  }
+
+  get neighborBombCount() {
+    return this.game.getNeighborBombCount(this.getRow(), this.getColumn())
   }
 
   getRow() {
@@ -895,15 +938,19 @@ class squareComponent extends AbstractTreeComponent {
     return this.getIndex()
   }
 
+  get game() {
+    return this.getRootNode().getGame()
+  }
+
   getCssClassNames() {
-    const game = this.getRootNode().getGame()
+    const game = this.game
     const row = this.getRow()
     const col = this.getColumn()
-    const wasClicked = game.wasClicked(row, col)
+    const wasClicked = this.wasClicked
     const isLost = game.isLost()
     const shouldReveal = game.shouldReveal()
     const neighborBombCount = game.getNeighborBombCount(row, col)
-    const isFlagged = game.isFlagged(row, col)
+    const isFlagged = this.isFlagged
     const hasBomb = game.hasBomb(row, col)
 
     let classNames = "squareComponent "
@@ -924,7 +971,8 @@ class squareComponent extends AbstractTreeComponent {
   }
 }
 
-class controlsComponent extends AbstractTreeComponent {
+// todo: STATE
+class controlsComponent extends AbstractSweeperCraftComponent {
   getStumpCode() {
     const parts = []
     const game = this.getRootNode().getGame()
@@ -938,19 +986,36 @@ class controlsComponent extends AbstractTreeComponent {
 
     return parts.join("\n") || "div"
   }
+
+  _syncBoardToGame() {
+    const game = this.getRootNode().getGame()
+    this.setContent(`${game.isOver() ? "over" : "notOver"} ${game.isFlagLockOn() ? "flagLockOn" : "flagLockOff"}`)
+  }
+}
+
+// todo: STATE
+class customLinkComponent extends AbstractSweeperCraftComponent {
+  getStumpCode() {
+    const craftLink = this._getGameLink()
+    if (craftLink) return `div Your game link: <a href="#${craftLink}">${craftLink}</a>`
+    return `div`
+  }
+
+  private _getGameLink() {
+    const game = this.getRootNode().getGame()
+    if (game.getNumberOfFlags() && !game.getNumberOfMoves()) return game.getCraftPermalink()
+    return ""
+  }
+
+  _syncBoardToGame() {
+    this.setContent(`${this._getGameLink()}`)
+  }
 }
 
 class shortcutsTableComponent extends AbstractTreeComponent {
   getStumpCode() {
-    const game = this.getRootNode().getGame()
-    let craftDiv = ""
-    if (game.getNumberOfFlags() && !game.getNumberOfMoves()) {
-      const craftLink = game.getCraftPermalink()
-      craftDiv = `Your game link: <a href="#${craftLink}">${craftLink}</a>`
-    }
     return `div
  id shortcuts
- div ${craftDiv}
  table
   tbody
    tr